Introduce benchmarks for Original Bloom
This commit introduces benchmarks for insertion and creation on the
original Bloom filter. Technically insertion is _also_ benchmarking
creation but I thought it'd be helpful to split these out.

I did a very little bit of optimization work from the benchmarks,
switching the underlying bitvec from a u8 field to u64. Non-trivial
boost, there: about 10% on the insertion numbers. Of course, my
machine is 64-bit.

+++ src/filters/bloom/original.rs
@@ -45,7 +45,7 @@ where
     capacity: usize,
     len: usize,
     hash_factors: Vec<usize>,
-    data: BitVec<BigEndian, u8>,
+    data: BitVec<BigEndian, u64>,
     phantom: PhantomData<K>,
 }
